addressed code review comments + fixed tests - to do - investigate cpu usage
