+++ .travis.yml
@@ -9,3 +9,4 @@ matrix:
       - go: tip
 script:
   go test -v -race -cpu=1,2,4 -bench . -benchmem ./...
+  go test -v -tags enable_binary_log -race -cpu=1,2,4 -bench . -benchmem ./...
