+++ array_test.go
@@ -1,5 +1,3 @@
-// +build !enable_binary_log
-
 package zerolog
 
 import (
@@ -26,7 +24,7 @@ func TestArray(t *testing.T) {
 		Time(time.Time{}).
 		Dur(0)
 	want := `[true,1,2,3,4,5,6,7,8,9,10,11,12,"a","0001-01-01T00:00:00Z",0]`
-	if got := string(a.write([]byte{})); got != want {
+	if got := decodeObjectToStr(a.write([]byte{})); got != want {
 		t.Errorf("Array.write()\ngot:  %s\nwant: %s", got, want)
 	}
 }
