+++ log/log_example_test.go
@@ -1,3 +1,5 @@
+// +build !enable_binary_log
+
 package log_test
 
 import (
