+++ syslog.go
@@ -1,4 +1,5 @@
 // +build !windows
+// +build !enable_binary_log
 
 package zerolog
 
@@ -28,13 +29,11 @@ func SyslogLevelWriter(w SyslogWriter) LevelWriter {
 }
 
 func (sw syslogWriter) Write(p []byte) (n int, err error) {
-	p = DecodeIfBinaryToBytes(p, true)
 	return sw.w.Write(p)
 }
 
 // WriteLevel implements LevelWriter interface.
 func (sw syslogWriter) WriteLevel(level Level, p []byte) (n int, err error) {
-	p = DecodeIfBinaryToBytes(p, true)
 	switch level {
 	case DebugLevel:
 		err = sw.w.Debug(string(p))
